Created a camera controller
The camera controller can move the camera around the player while respecting collisions so it will move closer to the player if there is an obstacle in the way. It also calculates the direction which is forward from the camera through the player so that a player controller can react to the movements of the camera.